Moved getPathNameInLocal8bit inside FileSystem namespace
