+++ launcher/InstanceImportTask.h
@@ -47,9 +47,6 @@
 #include <optional>
 
 class QuaZip;
-namespace Flame {
-class FileResolvingTask;
-}
 
 class InstanceImportTask : public InstanceTask {
     Q_OBJECT
@@ -79,7 +76,6 @@ class InstanceImportTask : public InstanceTask {
 
    private: /* data */
     NetJob::Ptr m_filesNetJob;
-    shared_qobject_ptr<Flame::FileResolvingTask> m_modIdResolver;
     QUrl m_sourceUrl;
     QString m_archivePath;
     bool m_downloadRequired = false;
